Debug asynchronous sockets error
The initial websocket handshake was being called outside the success
function of the AJAX call that receives the current user's id--due to
this asychronicity, sometimes the handshake was before the AJAX call
finished, meaning a proper Group socket connection was never made in
Django, and accordingly messages were not broadcasted to the user.

Fix this by putting the the initial handshake socket connection within
the AJAX call.